consolidate firebase handler
doing this because spawning multiple `app`s was creating problems
